requirements.txt
customtkinter
pillow
pandas
openpyxl
gdown
flask